requirements.txt
Django==1.11.14
djangorestframework==3.8.2
mysqlclient==1.3.13
Pillow==5.2.0
python-decouple==3.1
pytz==2018.5
reportlab==3.2
